backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
python-dotenv
sqlalchemy
pymysql
psycopg2-binary
